See merge request !16

+++ src/Listener/SaveReactionsToDatabase.php
@@ -1,21 +1,25 @@
 <?php
 
-/*
- * This file is part of datitisev/flarum-ext-reactions
+/**
+ *  This file is part of reflar/reactions
  *
- * © David Sevilla Martín <dsevilla192@icloud.com>
+ *  Copyright (c) ReFlar.
  *
- * For the full copyright and license information, please view the MIT license
+ *  http://reflar.io
+ *
+ *  For the full copyright and license information, please view the license.md
+ *  file that was distributed with this source code.
  */
 
-namespace Datitisev\Reactions\Listener;
+namespace Reflar\Reactions\Listener;
 
-use Datitisev\Reactions\Event\PostWasReacted;
-use Datitisev\Reactions\Event\PostWasUnreacted;
 use Flarum\Core\Access\AssertPermissionTrait;
 use Flarum\Event\PostWasDeleted;
 use Flarum\Event\PostWillBeSaved;
 use Illuminate\Contracts\Events\Dispatcher;
+use Reflar\Reactions\Event\PostWasReacted;
+use Reflar\Reactions\Event\PostWasUnreacted;
+use Reflar\Reactions\Reaction;
 
 class SaveReactionsToDatabase
 {
@@ -40,19 +44,23 @@ class SaveReactionsToDatabase
 
         if ($post->exists && isset($data['attributes']['reaction'])) {
             $actor = $event->actor;
-            $reacted = (bool) $data['attributes']['reaction'];
-            $reaction = $data['attributes']['reaction'];
+            $reacted = (bool)$data['attributes']['reaction'];
+            $reactionType = $data['attributes']['reaction'];
 
             $this->assertCan($actor, 'react', $post);
 
             $currentlyReacted = $post->reactions()->where('user_id', $actor->id)->exists();
 
-            if ($reacted && ! $currentlyReacted) {
-                $post->reactions()->attach($actor->id, ['reaction' => $reaction]);
+            if ($reacted && !$currentlyReacted) {
+                $reaction = Reaction::where('identifier', $reactionType)->first();
 
-                $post->raise(new PostWasReacted($post, $actor, $reaction));
+                $post->reactions()->attach($reaction, ['user_id' => $actor->id, 'reaction_id' => $reaction->identifier]);
+
+                $post->raise(new PostWasReacted($post, $actor, $reaction->identifier));
+
             } elseif ($currentlyReacted) {
-                $post->reactions()->detach($actor->id);
+
+                $post->reactions()->detach($post->reactions()->where('user_id', $actor->id)->first());
 
                 $post->raise(new PostWasUnreacted($post, $actor));
             }
